Split genericManifest implementations into image/docker_schema[12].go
... and their creation from getParsedManifest to
manifestSchema[12]FromManifest.

This will separate the top-level concerns from deailing with specific
manifest formats, and hopefully make the structure clearer.

Also reordered genericImage.Inspect to make it clear how it is similar
to genericImage.BlobDigests.

Does not change behavior.
